INS-2123: separate slot from conveyor
