+++ conveyor/conveyor_test.go
@@ -20,7 +20,6 @@ import (
 	"errors"
 	"fmt"
 	"testing"
-	"time"
 
 	"github.com/insolar/insolar/conveyor/queue"
 	"github.com/insolar/insolar/insolar"
@@ -103,8 +102,8 @@ func testPulseConveyor(t *testing.T, isQueueOk bool) *PulseConveyor {
 	return &PulseConveyor{
 		state:              insolar.ConveyorActive,
 		slotMap:            slotMap,
-		futurePulseNumber:  &presentPulse,
-		presentPulseNumber: &futurePulse,
+		futurePulseNumber:  &futurePulse,
+		presentPulseNumber: &presentPulse,
 	}
 }
 
@@ -368,19 +367,19 @@ func TestConveyor_ActivatePulse_NoPrepare(t *testing.T) {
 	require.Equal(t, insolar.ConveyorActive, c.state)
 }
 
-// func TestConveyor_ActivatePulse_PushSignalErr(t *testing.T) {
-// 	c := testPulseConveyor(t, false)
-// 	pulse := insolar.Pulse{PulseNumber: testRealPulse + testPulseDelta}
-// 	c.futurePulseData = &pulse
-// 	newFutureSlot := NewWorkingSlot(Future, pulse.NextPulseNumber, nil)
-// 	c.slotMap[pulse.NextPulseNumber] = newFutureSlot
-// 	c.state = insolar.ConveyorPreparingPulse
-//
-// 	panicValue := fmt.Sprintf("[ ActivatePulse ] can't send signal to future slot (for pulse %d), error - test error", c.futurePulseNumber)
-// 	require.PanicsWithValue(t, panicValue, func() { c.ActivatePulse() })
-// 	require.NotNil(t, c.futurePulseData)
-// 	require.Equal(t, insolar.ConveyorPreparingPulse, c.state)
-// }
+func TestConveyor_ActivatePulse_PushSignalErr(t *testing.T) {
+	c := testPulseConveyor(t, false)
+	pulse := insolar.Pulse{PulseNumber: testRealPulse + testPulseDelta}
+	c.futurePulseData = &pulse
+	newFutureSlot := mockSlot(t, false)
+	c.slotMap[pulse.NextPulseNumber] = newFutureSlot
+	c.state = insolar.ConveyorPreparingPulse
+
+	panicValue := fmt.Sprintf("[ ActivatePulse ] can't send signal to future slot (for pulse %d), error - test error", c.futurePulseNumber)
+	require.PanicsWithValue(t, panicValue, func() { c.ActivatePulse() })
+	require.NotNil(t, c.futurePulseData)
+	require.Equal(t, insolar.ConveyorPreparingPulse, c.state)
+}
 
 func TestConveyor_ActivatePreparePulse(t *testing.T) {
 	c := testPulseConveyor(t, true)
@@ -394,96 +393,3 @@ func TestConveyor_ActivatePreparePulse(t *testing.T) {
 	err = c.ActivatePulse()
 	require.NoError(t, err)
 }
-
-// ---- integration  tests
-
-func TestConveyor_ChangePulse(t *testing.T) {
-	conveyor, err := NewPulseConveyor()
-	require.NoError(t, err)
-	callback := mockCallback()
-	pulse := insolar.Pulse{PulseNumber: testRealPulse + testPulseDelta}
-	err = conveyor.PreparePulse(pulse, callback)
-	require.NoError(t, err)
-
-	callback.(*mockSyncDone).GetResult()
-
-	err = conveyor.ActivatePulse()
-	require.NoError(t, err)
-}
-
-func TestConveyor_ChangePulseMultipleTimes(t *testing.T) {
-	conveyor, err := NewPulseConveyor()
-	require.NoError(t, err)
-
-	pulseNumber := testRealPulse + testPulseDelta
-	for i := 0; i < 20; i++ {
-		callback := mockCallback()
-		pulseNumber += testPulseDelta
-		pulse := insolar.Pulse{PulseNumber: pulseNumber, NextPulseNumber: pulseNumber + testPulseDelta}
-		err = conveyor.PreparePulse(pulse, callback)
-		require.NoError(t, err)
-
-		callback.(*mockSyncDone).GetResult()
-
-		err = conveyor.ActivatePulse()
-		require.NoError(t, err)
-	}
-}
-
-func TestConveyor_ChangePulseMultipleTimes_WithEvents(t *testing.T) {
-	conveyor, err := NewPulseConveyor()
-	require.NoError(t, err)
-
-	pulseNumber := testRealPulse + testPulseDelta
-	for i := 0; i < 100; i++ {
-
-		go func() {
-			for j := 0; j < 1; j++ {
-				conveyor.SinkPush(pulseNumber, "TEST")
-				conveyor.SinkPush(pulseNumber-testPulseDelta, "TEST")
-				conveyor.SinkPush(pulseNumber+testPulseDelta, "TEST")
-				conveyor.SinkPushAll(pulseNumber, []interface{}{"TEST", i * j})
-			}
-		}()
-
-		go func() {
-			for j := 0; j < 100; j++ {
-				conveyor.GetState()
-			}
-		}()
-
-		go func() {
-			for j := 0; j < 100; j++ {
-				conveyor.IsOperational()
-			}
-		}()
-
-		callback := mockCallback()
-		pulseNumber += testPulseDelta
-		pulse := insolar.Pulse{PulseNumber: pulseNumber, NextPulseNumber: pulseNumber + testPulseDelta}
-		err = conveyor.PreparePulse(pulse, callback)
-		require.NoError(t, err)
-
-		if i == 0 {
-			require.Equal(t, 0, callback.(*mockSyncDone).GetResult())
-		} else {
-			require.Equal(t, 555, callback.(*mockSyncDone).GetResult())
-		}
-
-		err = conveyor.ActivatePulse()
-		require.NoError(t, err)
-
-		go func() {
-			for j := 0; j < 10; j++ {
-				require.NoError(t, conveyor.SinkPushAll(pulseNumber, []interface{}{"TEST", i}))
-				require.NoError(t, conveyor.SinkPush(pulseNumber, "TEST"))
-				require.NoError(t, conveyor.SinkPush(pulseNumber-testPulseDelta, "TEST"))
-				conveyor.SinkPush(pulseNumber+testPulseDelta, "TEST")
-			}
-		}()
-	}
-
-	time.Sleep(time.Millisecond * 200)
-}
-
-// TODO: Add test on InitiateShutdown
